refactor: move callout plugin into src dir
